add aggregator for shiny app
As Athena's fix remains uncertain both in timings and in expected performance improvements, I'd like to implement this scraper, which (through the scrape_all function) pulls together all the data in the desired format. Even if Athena improves enough for it to be practical to hook the shiny app directly to it, I think this is still a worthwhile interim step as it's easier to adapt the code.